Updated installer scripts for new stable rocm pytorch builds

+++ JTP3_HYDRA/Run ROCm.cmd
@@ -1,9 +1,9 @@
 @echo off
+title JTP³ Hydra ROCm
+set TORCH_ROCM_AOTRITON_ENABLE_EXPERIMENTAL=1
 if not exist venv python -m venv venv
 venv\Scripts\python -m pip install --upgrade pip
-set TORCH_ROCM_AOTRITON_ENABLE_EXPERIMENTAL=1
-AMDDetect.ps1
-title JTP³ Hydra ROCm
-venv\Scripts\pip install -r requirements.txt
+if not exist venv\Lib\site-packages\torch AMDDetect.ps1
+if not exist venv\Lib\site-packages\flask venv\Scripts\pip install -r requirements.txt
 cls
 venv\Scripts\python inference.py --service
