+++ JTP3_HYDRA/Run.cmd
@@ -2,5 +2,6 @@
 title JTP³ Hydra
 if not exist venv python -m venv venv
 venv\Scripts\python -m pip install --upgrade pip
-venv\Scripts\pip install -r requirements.txt
+if not exist venv\Lib\site-packages\torch venv\Scripts\pip install -r requirements.txt
+cls
 venv\Scripts\python inference.py --service
